:computer: Refactors Score Panel to it's own component
