+++ lib/dj_rumble_web/live/components/player_controls/score_panel/score_panel.html.leex
@@ -1,0 +1,20 @@
+<div class="grid grid-cols-4 items-center justify-center h-full p-6">
+
+  <div class="col-span-1 flex justify-end">
+    <%= render_score_button(:positive, @scoring_enabled.positive, @visitor, @myself) %>
+  </div>
+
+  <div class="col-span-2 grid grid-rows-2 justify-self-center">
+    <div class="row-span-1">
+      <span class="text-gray-400 text-5xl font-street-ruler">Live Score</span>
+    </div>
+    <div class="row-span-1">
+      <span class="text-gray-400 text-5xl font-street-ruler"><%= @live_score %></span>
+    </div>
+  </div>
+
+  <div class="col-span-1 flex justify-start">
+    <%= render_score_button(:negative, @scoring_enabled.negative, @visitor, @myself) %>
+  </div>
+
+</div>
